backend/requirements.txt
brainbase_labs
python-dotenv 
aiohttp